refactor(apps): move CLI adapter under apps
